pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "diane-cli"
version = "0.4.0"
description = "Externalized mental records clerk - frictionless thought capture CLI"
authors = [{name = "matias-ceau"}]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "python-dateutil>=2.8",
    "rich-click>=1.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "ruff>=0.1",
]
tui = [
    "textual>=0.40.0",
]
audio = [
    "openai>=1.0",
]
all = [
    "textual>=0.40.0",
    "openai>=1.0",
]

[project.scripts]
diane = "diane.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["diane*"]
exclude = ["tests*"]
